requirements.txt
fastapi
uvicorn
pandas
numpy
pillow
opencv-python-headless
